Move post write handling from StaticFiles to Core
